refactor: Move camelToKebab to utils and share the did-you-mean helper.
camelToKebab is needed by codegen, not just the plugin, and the Levenshtein
suggestion helper is about to serve a second kind of unknown name.
